Rename module to plural
